refactor(observe): atomic observe on Listeners; SignalMap + router delegate
Hoist the "select a derived value → `Object.is` dedup → optional immediate fire"
logic out of the per-consumer methods into a single `Listeners.observe(select,
listener, options?)` primitive (with `ObserveOptions` moved alongside it).

- `SignalMap.observe(key, …)` now delegates to `#listeners.observe(() =>
  this.get(key), …)`.
- `RouterController.observeCurrentActivity` delegates to `_listeners.observe(()
  => this.currentActivity, …)` — no more bespoke last-value bookkeeping.

`ObserveOptions` is re-exported from `di/SignalMap` so existing importers are
unaffected. Behavior-identical; full green gate (specs 1051, e2e 383, lint 0
errors).
